+++ src/components/ui/@react-table/components/table-column-filter.tsx
@@ -6,9 +6,9 @@ import { DateRange } from 'react-day-picker'
 import { useTranslation } from 'react-i18next'
 import { Div, DropdownSelect, Icon } from '../..'
 import { DateRangePicker } from '../../@core/date-range-picker'
+import { DebouncedInput, DebouncedInputProps } from '../../@custom/debounced-input'
 import { DEFAULT_ESTIMATE_SIZE } from '../constants'
 import { useTableContext } from '../context/table.context'
-import { DebouncedInput, DebouncedInputProps } from './debounced-input'
 import MultiSelectColumnFilter from './multi-select-column-filter'
 import { NumberRangeFilter } from './number-range-filter'
